# eds/api/api/models/humidity_summary.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class HumiditySummary(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, nm=None, u=None, v=None, t=None, n=None, lbl=None):
        """
        HumiditySummary - a model defined in Swagger

        :param nm: The nm of this HumiditySummary.
        :type nm: str
        :param u: The u of this HumiditySummary.
        :type u: str
        :param v: The v of this HumiditySummary.
        :type v: float
        :param t: The t of this HumiditySummary.
        :type t: datetime
        :param n: The n of this HumiditySummary.
        :type n: str
        :param lbl: The lbl of this HumiditySummary.
        :type lbl: str
        """
        self.swagger_types = {
            'nm': str,
            'u': str,
            'v': float,
            't': datetime,
            'n': str,
            'lbl': str
        }

        self.attribute_map = {
            'nm': 'nm',
            'u': 'u',
            'v': 'v',
            't': 't',
            'n': 'n',
            'lbl': 'lbl'
        }

        self._nm = nm
        self._u = u
        self._v = v
        self._t = t
        self._n = n
        self._lbl = lbl

    @classmethod
    def from_dict(cls, dikt):
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The HumiditySummary of this HumiditySummary.
        :rtype: HumiditySummary
        """
        return deserialize_model(dikt, cls)

    @property
    def nm(self):
        """
        Gets the nm of this HumiditySummary.

        :return: The nm of this HumiditySummary.
        :rtype: str
        """
        return self._nm

    @nm.setter
    def nm(self, nm):
        """
        Sets the nm of this HumiditySummary.

        :param nm: The nm of this HumiditySummary.
        :type nm: str
        """

        self._nm = nm

    @property
    def u(self):
        """
        Gets the u of this HumiditySummary.

        :return: The u of this HumiditySummary.
        :rtype: str
        """
        return self._u

    @u.setter
    def u(self, u):
        """
        Sets the u of this HumiditySummary.

        :param u: The u of this HumiditySummary.
        :type u: str
        """

        self._u = u

    @property
    def v(self):
        """
        Gets the v of this HumiditySummary.

        :return: The v of this HumiditySummary.
        :rtype: float
        """
        return self._v

    @v.setter
    def v(self, v):
        """
        Sets the v of this HumiditySummary.

        :param v: The v of this HumiditySummary.
        :type v: float
        """

        self._v = v

    @property
    def t(self):
        """
        Gets the t of this HumiditySummary.
        the timestamp when the humidity was measured

        :return: The t of this HumiditySummary.
        :rtype: datetime
        """
        return self._t

    @t.setter
    def t(self, t):
        """
        Sets the t of this HumiditySummary.
        the timestamp when the humidity was measured

        :param t: The t of this HumiditySummary.
        :type t: datetime
        """

        self._t = t

    @property
    def n(self):
        """
        Gets the n of this HumiditySummary.

        :return: The n of this HumiditySummary.
        :rtype: str
        """
        return self._n

    @n.setter
    def n(self, n):
        """
        Sets the n of this HumiditySummary.

        :param n: The n of this HumiditySummary.
        :type n: str
        """

        self._n = n

    @property
    def lbl(self):
        """
        Gets the lbl of this HumiditySummary.

        :return: The lbl of this HumiditySummary.
        :rtype: str
        """
        return self._lbl

    @lbl.setter
    def lbl(self, lbl):
        """
        Sets the lbl of this HumiditySummary.

        :param lbl: The lbl of this HumiditySummary.
        :type lbl: str
        """

        self._lbl = lbl